requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
pymysql==1.0.2
pillow==9.5.0
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.1.0
python-magic==0.4.27
tenacity==8.2.2

# ML/DL dependencies
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
opencv-python-headless==4.8.0.74  # Using headless version for better compatibility
timm==0.9.2
efficientnet-pytorch==0.7.1


# NLP dependencies
spacy==3.6.0
spacy-transformers==1.2.5
thefuzz==0.19.0
python-Levenshtein==0.21.1

# Database and search
faiss-cpu==1.7.4  # Use faiss-gpu for CUDA support
pandas==2.0.3

# Monitoring and visualization
wandb==0.15.4
matplotlib==3.7.1
seaborn==0.12.2
prometheus-client==0.17.1
tqdm==4.65.0

# Optional dependencies for enhanced features
scikit-image==0.21.0
albumentations==1.3.1

# Pre-trained models
clip @ git+https://github.com/openai/CLIP.git
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl